login,logout, register setup

+++ base/views.py
@@ -1,8 +1,15 @@
 from multiprocessing import context
+from django.http import HttpResponse
 from django.shortcuts import render,redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.forms import UserCreationForm
+
 # Create your views here.
 
 # rooms = [
@@ -11,6 +18,52 @@ from django.db.models import Q
 #     {'id': 3, 'name': 'Frontend developers'},
 # ]
 
+
+def loginPage(request):
+    page ='login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Username or password does not exist')
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or password does not exist.")
+            
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+    
+def logoutUser(request):
+    logout(request)
+    return redirect('home')    
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred during registration.")
+            
+    return render(request, 'base/login_register.html', {'form': form})
+    
 def home(request):
     # topic filter
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -34,6 +87,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -44,11 +98,15 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     #prefill form with room information
     form = RoomForm(instance=room)
+    
+    if request.user != room.user:
+        return HttpResponse('You are not allowed here.')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -59,9 +117,13 @@ def updateRoom(request, pk):
     context= {'form': form}
     return render(request, 'base/room_form.html', context)
     
-    
+@login_required(login_url='login')  
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.user:
+        return HttpResponse('You are not allowed here.')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
